requirements.txt
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Nuitka==2.7.9
ordered-set==4.1.0
packaging==25.0
pillow==11.2.1
pypdf==5.6.0
reportlab==4.4.2
Werkzeug==3.1.3
zstandard==0.23.0
qrcode
Pillow
pymupdf>=1.22.0
